+++ app/src/main/rust/hoshiepub/src/bin/uniffi-bindgen.rs
@@ -1,3 +1,4 @@
 fn main() {
-    uniffi::uniffi_bindgen_main()
+    uniffi::uniffi_bindgen_main();
 }
+